+++ changelog.md
@@ -17,6 +17,7 @@ This version is not yet released. If you are reading this on the website, then t
 - [`under ⍜`](https://uiua.org/docs/under) [`un °`](https://uiua.org/docs/un) [`bits ⋯`](https://uiua.org/docs/bits) now preserves a minimum bit length
 - [`under ⍜`](https://uiua.org/docs/under) [`reverse ⇌`](https://uiua.org/docs/reverse) and [`under ⍜`](https://uiua.org/docs/under) [`transpose ⍉`](https://uiua.org/docs/transpose) are now smarter with respect to the ranks of [`under ⍜`](https://uiua.org/docs/under)'s second function
   - For example `⍜⇌°⊂` behaves in a more useful way
+- Allow [`on ⟜`](https://uiua.org/docs/on) to use function packs
 - Allow [`under ⍜`](https://uiua.org/docs/under) of scalar [`keep ▽`](https://uiua.org/docs/keep)
 - Add the [`&fmd`](https://uiua.org/docs/&fmd) system function, which creates a directory
 - Add an experimental [`sort ⍆`](https://uiua.org/docs/sort) function